finishedProject/requirements.txt
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
decorator==4.4.2
ez-setup==0.9
google-api-core==2.8.2
google-auth==2.9.0
google-cloud-media-translation==0.8.2
googleapis-common-protos==1.56.3
grpcio==1.47.0
grpcio-status==1.47.0
idna==3.3
imageio==2.19.3
imageio-ffmpeg==0.4.7
moviepy==1.0.3
numpy==1.23.0
Pillow==9.2.0
proglog==0.1.10
proto-plus==1.20.6
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyPDF2==2.4.1
requests==2.28.1
rsa==4.8
six==1.16.0
soupsieve==2.3.2.post1
tk==0.1.0
tqdm==4.64.0
urllib3==1.26.9
